+++ package.json
@@ -28,6 +28,8 @@
     "eslint-plugin-jsx-a11y": "^1.2.0",
     "eslint-plugin-react": "^5.0.1",
     "http-server": "^0.9.0",
+    "ts-loader": "^0.8.2",
+    "typescript": "^1.9.0-dev.20160518-1.0",
     "webpack": "^1.13.0"
   },
   "eslintConfig": {
