+++ typings.json
@@ -1,0 +1,4 @@
+{
+  "name": "citeproc-js-docs",
+  "dependencies": {}
+}
